requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
alembic==1.7.4
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autopage==0.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cliff==3.9.0
cmaes==0.8.2
cmd2==2.2.0
colorama==0.4.4
colorlog==6.5.0
configparser==5.0.2
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.3.1
fsspec==2021.10.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
google-auth==2.3.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.0
huggingface-hub==0.0.19
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.0
jupyterlab-widgets==1.0.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.4
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pbr==5.6.0
pickleshare==0.7.5
pluggy==1.0.0
prettytable==2.2.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline3==3.3
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-lightning==1.4.9
pywin32==302
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.3.0
QtPy==1.11.2
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
sentencepiece==0.1.91
sentry-sdk==1.4.3
seqeval==1.2.2
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.26
stevedore==3.5.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.7.1+cu101
torchmetrics==0.5.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
yarl==1.7.0
yaspin==2.1.0
zipp==3.6.0
